Updated project to work with Visual Studio 2015. The changes are also reflected in the readme.

+++ dictionaryeditorform.cpp
@@ -138,7 +138,7 @@ QMimeData* DictionaryEditorModel::mimeData(const QModelIndexList &indexes) const
 
 void DictionaryEditorModel::handleAdded()
 {
-    signalRowsInserted({ { ZKanji::dictionaryCount() - 1, 1 } });
+    signalRowsInserted({ { (int)ZKanji::dictionaryCount() - 1, 1 } });
 }
 
 void DictionaryEditorModel::handleRemoved(int index, int order)
